requirements.txt
aniso8601==7.0.0
arrow==0.15.4
asgiref==3.2.3
blessings==1.7
bpython==0.18
certifi==2019.11.28
chardet==3.0.4
coreapi==2.3.3
coreschema==0.0.4
curtsies==0.3.0
dj-database-url==0.5.0
Django==3.0
django-rest-swagger==2.2.0
djangorestframework==3.11.0
graphene==2.1.8
graphene-django==2.7.1
graphql-core==2.2.1
graphql-relay==2.0.1
greenlet==0.4.15
gunicorn==20.0.4
idna==2.8
itypes==1.1.0
Jinja2==2.10.3
MarkupSafe==1.1.1
openapi-codec==1.3.2
Pillow==6.2.1
promise==2.3
Pygments==2.5.2
python-dateutil==2.8.1
pytz==2019.3
PyYAML==5.2
requests==2.22.0
Rx==1.6.1
simplejson==3.17.0
singledispatch==3.4.0.3
six==1.13.0
sqlparse==0.3.0
typing==3.7.4.1
uritemplate==3.0.0
urllib3==1.25.7
wcwidth==0.1.7
whitenoise==5.0.1
